requirements.txt
Pillow>=10.0.0
click>=8.1.7
pathlib>=1.0.1 